Added a mechanism to ignore non-certified test pbs
