+++ Packs/Logzio/pack_metadata.json
@@ -2,7 +2,7 @@
     "name": "Logz.io",
     "description": "Logz.io Package to fetch alerts from logz.io and threat hunting",
     "support": "partner",
-    "currentVersion": "1.1.0",
+    "currentVersion": "1.1.1",
     "author": "Logz.io",
     "url": "",
     "email": "support@logz.io",
